Fix profile initial scroll jumps
