+++ app/javascript/src/views/FreelancerProfile/index.js
@@ -1,14 +1,15 @@
-import React, { useEffect } from "react";
+import React from "react";
 // Hooks
 import useLogoURL from "src/components/ApplicationProvider/useLogoURL";
 import { useParams, Switch, Route } from "react-router-dom";
+import useInitialScroll from "./useInitialScroll";
 import useViewer from "src/hooks/useViewer";
 import { useQuery } from "@apollo/client";
 // Components
 import NotFound, { isNotFound } from "../NotFound";
 import Loading from "src/components/Loading";
 import AboutSection from "./AboutSection";
-import { Box, useBreakpoint } from "@advisable/donut";
+import { Box } from "@advisable/donut";
 import MainProfile from "./MainProfile";
 import GuildProfile from "./GuildProfile";
 // Queries
@@ -18,24 +19,12 @@ function FreelancerProfile() {
   useLogoURL("https://advisable.com");
   const params = useParams();
   const viewer = useViewer();
-  const widescreen = useBreakpoint("sUp");
   const { loading, data, error } = useQuery(GET_PROFILE, {
     variables: {
       id: params.id,
     },
   });
-
-  useEffect(() => {
-    if (window.history) {
-      window.history.scrollRestoration = "manual"; // disable automatic scroll restoration
-    }
-    const coverImgElement = document.getElementById("cover-img-wrapper");
-    const offset = coverImgElement?.offsetTop;
-    const height = coverImgElement?.offsetHeight;
-    const HEADER = 58;
-    const scrollY = offset - HEADER + (height / 100) * 64;
-    widescreen && data && scrollY && window.scrollTo(0, scrollY);
-  }, [data, widescreen]);
+  useInitialScroll(data);
 
   if (loading) return <Loading />;
   if (isNotFound(error)) return <NotFound />;
@@ -46,7 +35,7 @@ function FreelancerProfile() {
     <Box
       pb="2xl"
       mx={["12px", "32px", "32px", "auto"]}
-      maxWidth={["100%", "100%", "100%", "960px"]}
+      maxWidth={{ _: "100%", l: "960px" }}
     >
       <AboutSection
         specialist={data.specialist}
